requirements.txt
# ─── LangChain / LangGraph ────────────────────────────────────────────────────
langgraph>=0.2.0
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-community>=0.3.0
langchain-chroma>=0.1.4

# ─── LLM & Embeddings ─────────────────────────────────────────────────────────
openai>=1.40.0

# ─── Parsing documents ────────────────────────────────────────────────────────
pdfplumber>=0.11.0          # extraction texte PDF
pandas>=2.1.0               # parsing FEC (CSV tabulé)
openpyxl>=3.1.0             # lecture Excel si besoin

# ─── Vectorstore (RAG catalogue missions) ─────────────────────────────────────
chromadb>=0.5.0

# ─── UI ───────────────────────────────────────────────────────────────────────
streamlit>=1.38.0
plotly>=5.22.0              # graphiques radar, bar

# ─── Export Word ──────────────────────────────────────────────────────────────
python-docx>=1.1.0

# ─── HTTP / API BdF & INSEE ───────────────────────────────────────────────────
httpx>=0.27.0

# ─── Validation modèles ───────────────────────────────────────────────────────
pydantic>=2.7.0

# ─── Utilitaires ──────────────────────────────────────────────────────────────
python-dotenv>=1.0.0        # chargement .env
tenacity>=8.3.0             # retry sur appels API

# ─── Dev / Tests (optionnel) ──────────────────────────────────────────────────
pytest>=8.0.0
pytest-asyncio>=0.23.0
